corrected typo -- cart_itmes != cart_items

+++ db/carts.js
@@ -172,17 +172,10 @@ async function getAllPurchasedCarts() {
 
 async function deleteAbandonedGuestCarts() {
   try {
-    await client.query(
-      `DELETE FROM cart_items WHERE id IN 
-        (
-          SELECT cart_items.id as ccid FROM
-          cart_itmes JOIN carts ON
-          carts.id = cart_items.cart_id
-          WHERE purchased = false and
-          user_id = 9999
-        );
+    await client.query( //getting realtion "cart_items" does not exist
+      `DELETE FROM cart_items WHERE id IN (SELECT cart_items.id as ccid FROM cart_items JOIN carts ON carts.id = cart_items.cart_id WHERE purchased = FALSE AND user_id = 9999);
        DELETE FROM carts WHERE
-       purchased = false AND
+       purchased = FALSE AND
        user_id = 9999;
       `
     )
